\name{filterDEG_FDR}
\alias{filterDEG_FDR}
\title{
Filter FDR and plot DEG results
}
\description{
Function adapted from systemPipeR.
Filters and plots DEG results for a given set of sample comparisons. The gene idenifiers of all (i) Up_or_Down regulated genes are stored as separate list components and the corresponding summary statistics, stored in a one list component, is plotted in form of a stacked bar plot.
}
\usage{
filterDEG_FDR(degDF, filter, plot = TRUE, method)
}
\arguments{
  \item{degDF}{
  \code{data.frame} generated by \code{run_BaySeq}
}
  \item{filter}{
  Named vector with filter cutoffs of format \code{c(FDR=1)} where \code{FDR} to the p-value cutoff.
}
  \item{plot}{
  Allows to turn plotting behavior on and off with default set to \code{TRUE}. 
}
 \item{method}{
  Defines the method name in the plot.
}
}
\value{
Returns \code{list} with one components
  \item{UporDown}{List of up or down regulated gene/transcript indentifiers meeting the chosen filter settings for all comparisons defined in data frames \code{pval}.}
}
\references{
Thomas Girke (2015). systemPipeR: systemPipeR: NGS workflow and report generation environment. R package version 1.0.12.
https://github.com/tgirke/systemPipeR
}
\author{
Daniela Cassol
}
\seealso{
\code{run_BaySeq}
}
\keyword{run_DEGs}

