\name{panel.cor}
\alias{panel.cor}
\title{
panel.cor - Scatterplot 
}
\description{
panel.cor puts correlation in upper panels, size proportional to correlation.
}
\usage{
panel.cor(x, y, digits = 2, prefix = "", cex.cor, ...)
}
\value{
Returns \code{plot} with scatterplot matrix.
}
\references{
http://www.gettinggeneticsdone.com/2011/07/scatterplot-matrices-in-r.html
}
\author{
Daniela Cassol
}
\seealso{
\code{pairs}
}
\keyword{panel.cor}
