\name{filterDEG_FDR}
\alias{filterDEG_FDR}
\title{
Filter FDR and plot DEG results
}
\description{
Function adapted from systemPipeR.
Filters and plots DEG results for a given set of sample comparisons. The gene idenifiers of all (i) Up_or_Down regulated genes are stored as separate list components and the corresponding summary statistics, stored in a one list component, is plotted in form of a stacked bar plot.
}
\usage{
filterDEG_FDR(degDF, filter, plot = TRUE, method)
}
\arguments{
  \item{degDF}{
  \code{data.frame} generated by \code{run_BaySeq}
}
  \item{filter}{
  Named vector with filter cutoffs of format \code{c(FDR=1)} where \code{FDR} to the p-value cutoff.
}
  \item{plot}{
  Allows to turn plotting behavior on and off with default set to \code{TRUE}. 
}
 \item{method}{
  Defines the method name in the plot.
}
}
\value{
Returns \code{list} with one components
  \item{UporDown}{List of up or down regulated gene/transcript indentifiers meeting the chosen filter settings for all comparisons defined in data frames \code{pval}.}
}
\references{
Thomas Girke (2015). systemPipeR: systemPipeR: NGS workflow and report generation environment. R package version 1.0.12.
https://github.com/tgirke/systemPipeR
}
\author{
Daniela Cassol
}
\seealso{
\code{run_BaySeq}
}
\examples{
targetspath <- system.file("extdata", "targets.txt", package="DEG.comparison")
targets <- read.delim(targetspath, comment="#")
cmp <- readComp(file=targetspath, format="matrix", delim="-")
countDFeByg <- system.file("extdata", "countDFeByg.xls", package="DEG.comparison")
countDFeByg <- read.delim(countDFeByg, row.names=1)
Comp3 <- list(AP1.4_AP1.67=c("AP1.4A","AP1.4B", "AP1.67A", "AP1.67B"), AP3.4_AP3.67=c("AP3.4A","AP3.4B", "AP3.67A" ,"AP3.67B"), AG.4_AG.67=c("AG.4A", "AG.4B", "AG.67A","AG.67B"))
bayseqDF <- run_BaySeq(countDFeByg, Comp3, number=27416)
DEG_list_bayseqDF <- filterDEG_FDR(degDF=bayseqDF, filter=c(FDR=1), method="BaySeq")
DEG_list_bayseqDF$Summary[1:4,]   
}
\keyword{run_DEGs}

